Print pointer instead of UUID for XptrType .rodata.
It's easier to read back in.  I was wary of serializing out pointers, but
one: this isn't global, it only needs to be unique per serialized file,
after which it's pretty much ignored; and two: Python uses internal
pointers as ID's, so why not me?

+++ src/assemble_priv.h
@@ -28,7 +28,7 @@
  * block) is managed by scope[].
  */
 struct as_frame_t {
-        int funcno;
+        long long funcno;
         struct buffer_t af_locals;
         int fp;
         struct buffer_t af_args;
